requirements.txt
e3nn==0.5.1
numba==0.54.0
numpy==1.19.5
open3d==0.9.0.0
plyfile==0.7.4
PyYAML==6.0
PyYAML==5.4.1
PyYAML==6.0
recon==0.0.1
scipy==1.7.1
setuptools==47.1.0
spconv_cu102==2.2.6
torch==1.9.1+cu111
torch_cluster==1.6.1
torchmetrics==0.11.4
tqdm==4.62.3
transformers==4.25.1
trimesh==3.9.32
